requirements.txt
absl-py==1.0.0
accelerate==0.25.0
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.3.0
aniso8601==9.0.1
anndata==0.8.0
antlr4-python3-runtime==4.8
anyio==3.5.0
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
astroid==2.9.3
async-timeout==4.0.2
attrs==20.3.0
autofaiss==2.15.8
babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.6.0
bleach==4.1.0
blessings==1.7
blinker==1.7.0
braceexpand==0.1.7
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.11
chex==0.1.5
click==8.1.7
clip-anytorch==2.5.2
clip-retrieval==2.37.0
coca-pytorch==0.0.6
colorama==0.4.4
colorlog==6.6.0
commonmark==0.9.1
configparser==5.2.0
cycler==0.11.0
dalle2-pytorch==1.1.0
dataclasses==0.6
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.4
dm-tree==0.1.7
docker-pycreds==0.4.0
einops-exts==0.0.3
einops==0.4.1
ema-pytorch==0.0.10
-e git+https://github.com/Veldrovive/embedding-dataset-reordering.git@main#egg=embedding_dataset_reordering
embedding-reader==1.5.1
entrypoints==0.3
etils==0.8.0
exifread-nocycle==3.0.1
faiss-cpu==1.7.4
filelock==3.4.2
findspark==2.0.1
fire==0.4.0
flask-cors==3.0.10
flask-restful==0.3.10
flask==2.3.3
flax==0.6.1
fonttools==4.29.0
frozenlist==1.3.0
fsspec==2022.11.0
ftfy==6.1.1
future==0.18.2
gdown==4.5.1
gitdb==4.0.9
gitpython==3.1.26
google-auth-oauthlib==0.4.6
google-auth==2.5.0
gpustat==0.6.0
grpcio==1.43.0
h5py==3.6.0
htmlmin==0.1.12
huggingface-hub==0.17.3
hydra-colorlog==1.1.0
hydra-core==1.1.1
hydra-joblib-launcher==1.1.5
identify==2.4.6
idna==3.3
imagehash==4.2.1
imageio==2.14.1
img2dataset==1.42.0
importlib-metadata==4.10.1
importlib-resources==5.4.0
iniconfig==1.1.1
ipdb==0.13.9
ipykernel==6.7.0
ipython-genutils==0.2.0
ipython==7.31.1
ipywidgets==7.6.5
isort==5.10.1
itsdangerous==2.1.2
jammy==0.0.6
jax==0.4.21
jaxlib==0.4.21
jedi==0.18.1
jinja2==3.1.2
joblib==1.0.1
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.1
jupyter-packaging==0.12.3
jupyter-server==1.13.4
jupyter==1.0.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.2
jupyterlab==3.2.8
kiwisolver==1.3.2
kornia==0.5.11
lazy-object-proxy==1.7.1
line-profiler==3.4.0
llvmlite==0.39.1
lmdb==1.3.0
loguru==0.5.3
lpips==0.1.4
markdown==3.3.6
markupsafe==2.1.3
matplotlib-inline==0.1.3
matplotlib==3.5.1
mccabe==0.6.1
missingno==0.5.1
mistune==0.8.4
msgpack-numpy==0.4.7.1
msgpack==1.0.3
multidict==6.0.2
multilingual-clip==1.0.10
multimethod==1.7
mypy-extensions==0.4.3
natsort==8.2.0
nbclassic==0.3.5
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
ninja==1.10.2.4
nltk==3.8.1
nodeenv==1.6.0
notebook==6.4.8
numba==0.56.2
numpy==1.21.5
nvidia-ml-py3==7.352.0
nvidia-ml-py==11.495.46
nvitop==0.8.1
oauthlib==3.2.0
omegaconf==2.1.1
open-clip-torch==2.23.0
open3d==0.15.2
opencv-python-headless==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
optax==0.1.3
ordered-set==4.1.0
ott-jax==0.2.9
packaging==21.3
pandas-profiling==3.1.0
pandas==1.1.5
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
patsy==0.5.3
pexpect==4.8.0
phik==0.12.0
pickleshare==0.7.5
pillow==9.0.0
pip==23.3.1
platformdirs==2.4.1
plotext==5.2.8
pluggy==1.0.0
pot==0.8.2
pre-commit==2.17.0
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.26
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pudb==2022.1
py4j==0.10.9.7
py==1.11.0
pyarrow==6.0.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pydeprecate==0.3.1
pygeos==0.13
pygments==2.11.2
pylint==2.12.2
pynndescent==0.5.7
pyparsing==3.0.7
pyquaternion==0.9.9
pyrsistent==0.18.1
pysocks==1.7.1
pyspark==3.5.0
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
python-memcached==1.59
pytorch-lightning==1.5.2
pytorch-warmup==0.1.0
pytz==2021.3
pywavelets==1.2.0
pyyaml==6.0
pyzmq==22.3.0
qtconsole==5.2.2
qtpy==2.0.1
qudida==0.0.4
regex==2022.8.17
requests-oauthlib==1.3.1
requests==2.27.1
resize-right==0.0.2
rich==12.6.0
rotary-embedding-torch==0.1.5
rsa==4.8
safetensors==0.4.1
scanpy==1.9.1
scikit-image==0.18.3
scikit-learn==1.1.2
scipy==1.7.3
seaborn==0.11.2
send2trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.5.4
session-info==1.0.0
setproctitle==1.3.2
setuptools-scm==6.4.2
setuptools==59.8.0
sh==1.14.2
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
statsmodels==0.13.2
stdlib-list==0.8.0
subprocess32==3.5.4
tangled-up-in-unicode==0.1.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.10.0
termcolor==2.0.1
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.11.2
timm==0.9.12
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.0
tomlkit==0.11.5
toolz==0.12.0
torch-ema==0.3
torch-fidelity==0.3.0
torchinfo==1.6.3
torchmetrics==0.9.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.35.2
typing-extensions==4.4.0
ubelt==0.10.2
umap-learn==0.5.3
urllib3==1.25.11
urwid-readline==0.13
urwid==2.1.2
vector-quantize-pytorch==0.9.2
virtualenv==20.13.0
visions==0.7.4
wandb==0.12.21
wcwidth==0.2.5
webdataset==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
werkzeug==3.0.1
wheel==0.37.1
widgetsnbextension==3.5.2
wrapt==1.13.3
x-clip==0.9.0
yarl==1.7.2
zipp==3.7.0
